adapt Account Functions to a JWT tokens

+++ xiang-app/app/Http/Controllers/AccountController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Account;
 use Illuminate\Routing\Controller as BaseController;
+use Tymon\JWTAuth\Facades\JWTAuth;
+use Tymon\JWTAuth\Exceptions\JWTException;
+use \Illuminate\Http\JsonResponse;
 
 class AccountController extends BaseController
 {
@@ -13,16 +16,35 @@ class AccountController extends BaseController
     {
         $this->middleware('auth:api');
     }
+
+    // Helper to test with the token is valid
+    protected function getAuthenticatedUser()
+    {
+        try {
+            if (!$user = JWTAuth::parseToken()->authenticate()) {
+                return response()->json(['error' => 'User not found'], 401);
+            }
+            return $user;
+        } catch (JWTException $e) {
+            return response()->json(['error' => 'Token is invalid or expired'], 401);
+        }
+    }
     
     //List of accounts
     public function index()
     {
-        return Account::where('user_id', auth()->guard('api')->user()->id)->get();
+        $user = $this->getAuthenticatedUser();
+        if($user instanceof JsonResponse) return $user;
+
+        return Account::where('user_id', $user->id)->get();
     }
 
     //Storage a new account
     public function store(Request $request)
     {
+        $user = $this->getAuthenticatedUser();
+        if ($user instanceof \Illuminate\Http\JsonResponse) return $user;
+
         $data = $request->all();
         $data['user_id'] = auth()->guard('api')->user()->id;
         Account::create($data);
@@ -32,6 +54,9 @@ class AccountController extends BaseController
     //Display a specific account
     public function show(string $id )
     {
+        $user = $this->getAuthenticatedUser();
+        if ($user instanceof \Illuminate\Http\JsonResponse) return $user;
+
         return Account::where('id', $id)
             ->where('user_id',auth()->guard('api')->user()->id)
             ->firstOrFail();
@@ -40,6 +65,9 @@ class AccountController extends BaseController
     //Update a specific account
     public function update(Request $request, string $id)
     {
+        $user = $this->getAuthenticatedUser();
+        if ($user instanceof \Illuminate\Http\JsonResponse) return $user;
+
         $account = Account::where('id', $id)->where('user_id',auth()->guard('api')->user()->id)->firstOrFail();
         $account->update($request->all());
         return $account;
@@ -48,6 +76,9 @@ class AccountController extends BaseController
     //Delete a specific account
     public function destroy(string $id)
     {
+        $user = $this->getAuthenticatedUser();
+        if ($user instanceof \Illuminate\Http\JsonResponse) return $user;
+
         $account = Account::where('id', $id)->where('user_id', auth()->guard('api')->user()->id)->firstOrFail();
         $account->delete();
         return response()->json(['message' => 'Account deleted successfully'], 200);
